+++ packages/client/src/components/organisms/Game/GameConfig.ts
