+++ packages/client/src/components/organisms/Game/index.ts
@@ -1,0 +1,4 @@
+import { onFinishData } from './core/Types'
+import { Game } from './core/Game'
+
+export { Game, type onFinishData }
