+++ packages/client/src/components/organisms/Game/utils/CheckShipsOnBoard.ts
@@ -1,0 +1,5 @@
+import { cellType } from '../core/Types'
+
+export const CheckShipsOnBoard = (board: cellType[][]) => {
+  return board.some(el => el.some(el => el === 'ship'))
+}
