+++ packages/client/src/components/organisms/Game/utils/Math.ts
